Split the art from classes to its own file
