+++ fiefdomgame.py
@@ -7,6 +7,7 @@ import bcrypt
 from classes import *
 from worldmap import *
 from os.path import exists
+from art import *
 
 '''
